Fix the narrowing warnings that were appearing on windows tiles builds

+++ cata_tiles.cpp
@@ -860,10 +860,10 @@ bool cata_tiles::draw_furniture(int x, int y)
     // for rotation inforomation
     const int neighborhood[4] =
     {
-        g->m.furn(x, y+1), // south
-        g->m.furn(x+1, y), // east
-        g->m.furn(x-1, y), // west
-        g->m.furn(x, y-1)  // north
+        static_cast<int> (g->m.furn(x, y+1)), // south
+        static_cast<int> (g->m.furn(x+1, y)), // east
+        static_cast<int> (g->m.furn(x-1, y)), // west
+        static_cast<int> (g->m.furn(x, y-1))  // north
     };
 
     int subtile = 0, rotation = 0;
@@ -884,10 +884,10 @@ bool cata_tiles::draw_trap(int x, int y)
 
     const int neighborhood[4] =
     {
-        g->m.tr_at(x, y+1), // south
-        g->m.tr_at(x+1, y), // east
-        g->m.tr_at(x-1, y), // west
-        g->m.tr_at(x, y-1)  // north
+        static_cast<int> (g->m.tr_at(x, y+1)), // south
+        static_cast<int> (g->m.tr_at(x+1, y)), // east
+        static_cast<int> (g->m.tr_at(x-1, y)), // west
+        static_cast<int> (g->m.tr_at(x, y-1))  // north
     };
 
     int subtile = 0, rotation = 0;
@@ -911,10 +911,10 @@ bool cata_tiles::draw_field_or_item(int x, int y)
         // for rotation inforomation
         const int neighborhood[4] =
         {
-            g->m.field_at(x, y+1).fieldSymbol(), // south
-            g->m.field_at(x+1, y).fieldSymbol(), // east
-            g->m.field_at(x-1, y).fieldSymbol(), // west
-            g->m.field_at(x, y-1).fieldSymbol()  // north
+            static_cast<int> (g->m.field_at(x, y+1).fieldSymbol()), // south
+            static_cast<int> (g->m.field_at(x+1, y).fieldSymbol()), // east
+            static_cast<int> (g->m.field_at(x-1, y).fieldSymbol()), // west
+            static_cast<int> (g->m.field_at(x, y-1).fieldSymbol())  // north
         };
 
         int subtile = 0, rotation = 0;
